requirements.txt




## needs python 3.8 or more .. needs openai >= 1
pyautogen    # pip install pyautogen

## optional dependencies -- docker

## to read environment variables as key-value pairs

python-dotenv    ## pip install python-dotenv


## for LLM and chat LMs, will be using OpenAI


## for any vectorestore, will be using Milvus vectore DB
pymilvus ## pip install milvus


## installing langchain to use openai in general
langchain
langchain_experimental   ## in order to import experimental features e g create_csv_agent

#### serpapi -- pip install google-search-results
google-search-results

#### -- ModuleNotFoundError: No module named 'numexpr'
numexpr   ## error while running agent ... pip install numexpr


### pip install pandas
pandas ## needed for csv agent to work
tabulate  ## also needed by csv agent


## yfinance to ask finance related questions and build agents
yfinance # pip install yfinance


### need pyspark for spark agent type -- in order to create spark session
pyspark # pip install pyspark